+++ CHANGELOG.md
@@ -57,3 +57,4 @@
 - [2026-04-02 12:51] feat(navigation): implement responsive navigation view component
 - [2026-04-02 12:55] docs(ui): update readme and API contracts for ui
 - [2026-04-02 13:03] feat(cache): add bi-directional event bus handler for cache
+- [2026-04-02 13:13] docs(theme): update readme and API contracts for theme